Divide tests onto several classes
