+++ app/src/androidTest/java/ru/spbu/depnav/data/db/AppDatabaseDaoTest.kt
@@ -1,0 +1,46 @@
+/**
+ * DepNav -- department navigator.
+ * Copyright (C) 2022  Timofey Pushkin
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package ru.spbu.depnav.data.db
+
+import android.content.Context
+import androidx.room.Room
+import androidx.test.core.app.ApplicationProvider
+import org.junit.After
+import org.junit.Before
+
+/** Base class for instrumentation tests for [AppDatabase] DAOs. */
+abstract class AppDatabaseDaoTest {
+    protected lateinit var db: AppDatabase
+
+    /** Initializes an instance of the database. */
+    @Before
+    fun setUpAppDatabase() {
+        val context = ApplicationProvider.getApplicationContext<Context>()
+        db = Room.inMemoryDatabaseBuilder(context, AppDatabase::class.java).build()
+    }
+
+    /** Closes the database instance. */
+    @After
+    fun tearDownAppDatabase() {
+        db.close()
+    }
+
+    /** Initializes a DAO instance. */
+    abstract fun setUpDao()
+}
